requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.121.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
reportlab==4.2.5
python-multipart==0.0.9
